Change u64 to u32 to prevent JSON issues

+++ src/game.rs
@@ -5,12 +5,12 @@ use crate::grid::Grid;
 
 #[derive(Serialize)]
 pub(crate) struct Game {
-    id: u64,
+    id: u32,
     #[serde(skip_serializing)]
     grid: Grid,
 }
 impl Game {
-    pub fn new(size: (u32, u32), game_id_counter: &Arc<Mutex<u64>>) -> Self {
+    pub fn new(size: (u32, u32), game_id_counter: &Arc<Mutex<u32>>) -> Self {
         Game {
             id: *game_id_counter.lock().unwrap(),
             grid: Grid::new(size),
